Extract identificationInformation to config/pages
