+++ src/applications/simple-forms-form-engine/shared/config/pages/index.js
@@ -1,0 +1,4 @@
+export {
+  default as identificationInformation,
+} from './identificationInformation';
+export { default as nameAndDateOfBirth } from './nameAndDateOfBirth';
